ChallengerLibraryBookCatalog/requirements.txt
asgiref==3.4.1
attrs==21.2.0
backcall==0.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.1
charset-normalizer==2.0.6
colorama==0.4.4
cryptography==3.4.8
decorator==5.1.0
distlib==0.3.3
distro==1.6.0
Django==3.2.7
django-bootstrap4==3.0.1
django-haystack==3.1.1
django-modeltranslation==0.17.3
docker==5.0.2
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
elasticsearch==5.5.3
filelock==3.0.12
identify==2.2.15
idna==3.2
ipdb==0.13.9
ipython==7.27.0
jedi==0.18.0
jsonschema==3.2.0
matplotlib-inline==0.1.3
nodeenv==1.6.0
paramiko==2.7.2
parso==0.8.2
pickleshare==0.7.5
platformdirs==2.3.0
pre-commit==2.15.0
prompt-toolkit==3.0.20
pycparser==2.20
Pygments==2.10.0
PyNaCl==1.4.0
pyrsistent==0.18.0
python-dotenv==0.19.0
pytz==2021.1
pywin32==227
PyYAML==5.4.1
requests==2.26.0
rjsmin==1.1.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
texttable==1.6.4
toml==0.10.2
traitlets==5.1.0
urllib3==1.26.7
virtualenv==20.8.1
wcwidth==0.2.5
websocket-client==0.59.0